dtype 32

+++ datasets/ReefDataset.py
@@ -91,7 +91,7 @@ class ReefDataset(Dataset):
                                           bboxes=target['boxes'],
                                           class_labels=target['labels'])
             img = torch.as_tensor(transformed['image'])
-            target["boxes"] = torch.as_tensor(transformed['bboxes'])
+            target["boxes"] = torch.as_tensor(transformed['bboxes'], dtype=torch.float32)
             target["labels"] = torch.as_tensor(transformed['class_labels'])
         
         print(img)
